A few bits of AVRG: Remove explicit request for control, instead, one instance per AVRG per room is always automatically in control. Also implement flow granting through quest completion, and a first untested stab at the automatic flow payout factor algorithm. I chopped up some bits of GameRepository while I was at it since most of that stuff was very standalone-game-centric, and yet AVRGs need to access it.

+++ src/java/com/threerings/msoy/avrg/client/AVRGameService.java
@@ -33,7 +33,7 @@ public interface AVRGameService extends InvocationService
         Client caller, String questId, int step, String progress, ConfirmListener listener);
 
     public void completeQuest (
-        Client caller, String questId, int payoutLevel, ConfirmListener listener);
+        Client caller, String questId, float payoutLevel, ConfirmListener listener);
 
     public void cancelQuest (
         Client caller, String questId, ConfirmListener listener);
